Move ascii::parse::test module into separate file
